app/AI_Python/requirements.txt
brotli==1.0.9
cjm-pandas-utils==0.0.3
cjm-pytorch-utils==0.0.6
fqdn==1.5.1
isoduration==20.11.0
jsonpointer==2.4
matplotlib==3.9.0
mkl-fft==1.3.8
mkl-random==1.2.4
notebook==7.2.0
pillow-avif-plugin==1.4.3
pip-chill==1.0.3
pysocks==1.7.1
scikit-learn==1.4.2
torchaudio==2.3.0
torcheval==0.0.7
torchtnt==0.2.3
uri-template==1.3.0
webcolors==1.13
win-inet-pton==1.1.0